Create methods in Factory taking in Domain Object
The methods delegate to the method of the same name with all of the domain object's properties.

Example:
UserFactory has 2 createNew() methods:
- One taking in ID, version, username, password.
- One taking in a whole User object.
The second one calls the first one with all of the User object's properties.

See #24

+++ src/dom/model/user/mapper/UserInputMapper.java
@@ -87,7 +87,7 @@ public class UserInputMapper {
 		
 		return UserFactory.createClean(
 				rs.getLong("id"),
-				rs.getInt("version"),
+				rs.getLong("version"),
 				rs.getString("username"),
 				rs.getString("password")
 		);
